Update inputs - Working check for unique ID added
When adding or editing an employee record, a separate list of employee IDs is generated and stored as id_list. 

When the user enters a new employee ID, A check is made to id_list.
The new user entry succeeds if it doesn't match an ID in id_list OR if the entry matches the employee's current ID from its JSON string (entry["employee_id"]).